Handle closing inside string differently
If inside a string no pending opener is found look for a normal opener.
This fixes the common problem when calling on a closing string delimiter
like '"' does nothing.
However when we actually are inside a string and not on the last
delimiter this allows infinite closing.
Idea: insert closer and check if it's no longer a string.